# Optimize performance of transfer