--- a/mars/services/storage/transfer.py
+++ b/mars/services/storage/transfer.py
@@ -107,14 +107,9 @@ class SenderManagerActor(mo.StatelessActor):
 
             async def flush(self):
                 if self._buffers:
-                    transfer_message = TransferMessage(
-                        self._buffers,
-                        session_id,
-                        self._send_keys,
-                        level,
-                        self._eof_marks,
+                    await receiver_ref.receive_part_data(
+                        self._buffers, session_id, self._send_keys, self._eof_marks
                     )
-                    await receiver_ref.receive_part_data(transfer_message)
 
                 self._buffers = []
                 self._send_keys = []
@@ -317,19 +312,18 @@ class ReceiverManagerActor(mo.StatelessActor):
                 future.cancel()
                 raise
 
-    async def do_write(self, message: TransferMessage):
+    async def do_write(
+        self, data: list, session_id: str, data_keys: List[str], eof_marks: List[bool]
+    ):
         # close may be a high-cost operation, use create_task
         close_tasks = []
         finished_keys = []
-        session_id = message.session_id
-        for data, data_key, is_eof in zip(
-            message.data, message.data_keys, message.eof_marks
-        ):
+        for data, data_key, is_eof in zip(data, data_keys, eof_marks):
             writer = self._writing_infos[(session_id, data_key)].writer
             if data:
                 await writer.write(data)
             if is_eof:
-                close_tasks.append(asyncio.create_task(writer.close()))
+                close_tasks.append(writer.close())
                 finished_keys.append(data_key)
         await asyncio.gather(*close_tasks)
         async with self._lock:
@@ -338,25 +332,23 @@ class ReceiverManagerActor(mo.StatelessActor):
                 event.set()
                 self._decref_writing_key(session_id, data_key)
 
-    async def receive_part_data(self, message: TransferMessage):
-        write_task = asyncio.create_task(self.do_write(message))
+    async def receive_part_data(
+        self, data: list, session_id: str, data_keys: List[str], eof_marks: List[bool]
+    ):
+        write_task = asyncio.create_task(
+            self.do_write(data, session_id, data_keys, eof_marks)
+        )
         try:
             await asyncio.shield(write_task)
         except asyncio.CancelledError:
-            session_id = message.session_id
             async with self._lock:
-                for data_key in message.data_keys:
-                    if (message.session_id, data_key) in self._writing_infos:
-                        if (
-                            self._writing_infos[
-                                (message.session_id, data_key)
-                            ].ref_counts
-                            == 1
-                        ):
-                            info = self._writing_infos[(message.session_id, data_key)]
+                for data_key in data_keys:
+                    if (session_id, data_key) in self._writing_infos:
+                        if self._writing_infos[(session_id, data_key)].ref_counts == 1:
+                            info = self._writing_infos[(session_id, data_key)]
                             await self._quota_refs[info.level].release_quota(info.size)
                             await self._storage_handler.delete(
-                                message.session_id, data_key, error="ignore"
+                                session_id, data_key, error="ignore"
                             )
                             await info.writer.clean_up()
                             info.event.set()