@qinxuye 2022-05-28
The master cannot run storage asv_bench directly, so the ci will fail.